requirements.txt
streamlit==1.30.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
validators==0.22.0
protobuf>=3.20.0
pyairtable==2.1.0
